确保同样的int32的值通过$cast能够得到同样$date及$timestamp SEQUOIADBMAINSTREAM-1280
git-svn-id: http://192.168.20.11/sequoiadb/trunk@22905 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/mth/mthSActionFunc.cpp
@@ -1126,7 +1126,7 @@ namespace engine
          }
          else if ( Timestamp == e.type() )
          {
-            INT32 l = e.timestampTime().millis ;
+            UINT64 l = e.timestampTime().millis ;
             l += e.timestampInc() / 1000 ;
             builder.appendNumber( fieldName, ( INT32 )l ) ;
          }
